refactor(logging): move response-reporting helpers out of the database class
`describeResponse` and the status naming encode how remote monitoring groups
and titles an issue, not how the database talks to CouchDB, so they do not
belong in RemotePouchDatabase. Moved to core/logging alongside the grouping
code they exist for, with their own spec.

Also trims the status-name table from seven entries to the three that
plausibly occur and each call for a different response: 400 (malformed
request, e.g. a bad query or a proxy mangling the url), 413 (payload over the
reverse proxy's 25M limit, e.g. a large attachment) and 429 (throttling).
The rest share one bucket and still carry their numeric status in the context;
naming statuses that never occur only made the list look authoritative.

No behaviour change. The database spec keeps the routing checks and drops the
message-wording assertions now covered where the wording lives.
